Finished refactoring, extensive name changes in home.css
