--- client/src/components/BackgroundImageTransition.jsx
@@ -1,3 +1,0 @@
-export default function () {
-  return <div className="background-image-transition"></div>;
-}
