+++ client/src/components/Hero.jsx
@@ -1,0 +1,12 @@
+import QuoteDisplay from "../components/QuoteDisplay";
+import AccountButtons from "../components/AccountButtons";
+
+export default function () {
+  return (<>
+      <section className="hero-background">
+        <QuoteDisplay />
+        <AccountButtons />
+        <div className="hero-transition"></div>
+      </section>
+    </>);
+}
