+++ client/src/components/MoreMoviesButton.jsx
@@ -1,11 +1,9 @@
 import { Link } from "react-router-dom";
 
 export default function () {
-  return (
-    <>
+  return (<>
       <Link to="movies">
         <button className="button">MORE MOVIES</button>
       </Link>
-    </>
-  );
+    </>);
 }
