+++ client/src/components/QuoteDisplay.jsx
@@ -1,13 +1,13 @@
-import mirror3 from "../images/mirror3.png";
+import QuoteMirror from "../images/QuoteMirror.png";
 import Quotes from "./Quotes.jsx";
 
 export default function () {
   return (
-    <div className="mirror-container">
-      <img className="mirror" src={mirror3} alt="a picture of an old mirror" />
-      <h3 id="backgroundQuote" className={"transparent-background"}>
-        <Quotes />
-      </h3>
+    <div className="display-container">
+      <img className="mirror" src={QuoteMirror} alt="an old decorative mirror" />
+      <h2 id="quote-text-container" className={"quote-background"}>
+        <Quotes/>
+      </h2>
     </div>
   );
 }
