+++ client/src/pages/HomePage.jsx
@@ -1,18 +1,9 @@
-import "../styles/home.css";
-import QuoteDisplay from "../components/QuoteDisplay";
-import AccountButtons from "../components/AccountButtons";
-import BackgroundImageTransition from "../components/BackgroundImageTransition";
-import TopMovieArea from "../components/TopMovieArea";
+import Hero from "../components/Hero";
+import TopMovies from "../components/TopMovies";
 
 export default function () {
   return (<>
-      <section className="background-image">
-        <QuoteDisplay />
-        <AccountButtons />
-        <BackgroundImageTransition />
-      </section>
-      <section className="top-movies">
-        <TopMovieArea />
-      </section>
+      <Hero />
+      <TopMovies />
     </>);
 }
